Create useless minimal bootloader
Goal here was to create something which qemu will accept as a bootable
image. No useful functionality is present at the moment.

Intel syntax is more or less an arbitrary decision at this point. It's
not the default for *nix development, but it has less syntactic noise,
and is clearer in some respects. Once I get more experienced with
assembly (if ever), switching to AT&T syntax hopefully will be
relatively easy. Considering that I want to try to program as much of
the code as possible in a higher-level language, though, this may not
ever happen...

Had to use an unusual (at least compared to the methods used in the
various internet tutorials) method of placing the bytes 0x55 and 0xaa at
offsets 510 and 511 in the output binary (this is necessary to ensure
the BIOS recognizes the device as containing a valid bootloader). The
most common method I saw on the internet to accomplish this was to use
the "times" directive supported by NASM to fill the intervening space
with zeros:

    times 510 - ($ - $$) db 0x00
    db 0x55
    db 0xaa

While this works, I wanted to use Clang as my assembler, and Clang's
assembler doesn't support the "times" directive. The desire to use Clang
as my assembler was more or less an arbitrary choice, since as far as I
know for most common cases NASM would work just as well. It just felt
weird to resort to NASM when Clang should work perfectly fine for
assembly.

The equivalent Clang assembly might look like this:

    .fill 510 - (. - mbr_boot)
    .byte 0x55
    .byte 0xaa

Where mbr_boot is a label positioned somewhere prior to the first
instruction in the assembly file.

However, if a "jmp mbr_boot" instruction (conditional or unconditional)
were present at any point between the .fill directive and the mbr_boot
label, Clang would fail to compile the file, complaining that it
expected an absolute expression for the .fill directive:

    rp:nih awang$ make bootloader
    /usr/local/opt/llvm/bin/clang --target=i386-elf -m16 -c boot.s -o boot.o
    boot.s:12:7: error: expected absolute expression
    .fill 510 - (. - mbr_boot)
          ^
    make: *** [bootloader] Error 1

This seems to be a bug in LLVM. From the llvm-dev mailing list [0]:

> While trying to compile an existing codebase which uses handwritten
> assembly with LLVM, I ran into an issue around using backward references in
> assembly absolute expressions. A simple example can be the following
> snippet:
>
> _foo:
>   .fill 0x100
> _bar:
>   .fill _bar - _foo
> _baz:
>   .fill 0x100
>
> While gas compiles this snippet without any errors, the integrated
> assembler throws an error: expected absolute expression for _bar - _foo.
>
> I haven't found any definition of absolute expression in gas manual, and
> it's arguable whether this case should be considered an absolute
> expression, but at the point of evaluating the directive, the addresses of
> backward labels should be already known so this expression could be treated
> as absolute. A quick search also revealed several bug reports related to
> this issue so it seems like this use case is fairly common.

It appears that GCC uses /usr/bin/as to compile assembly files on macOS,
so switching compilers wouldn't help here (and it seems that brew
doesn't compile/install gas). "jmp" is a fairly useful instruction, too,
so using .fill for padding isn't a viable option.

Luckily, it turns out that I can take advantage of the way the "org"
directive differs between Clang's assembler and NASM. The "org"
directive is supposed to denote where the assembler should expect the
code to be loaded (normally 0x0000:0x7C00 or 0x07C0:0x0000). With NASM,
the directive didn't appear to change the output file beyond tweaking
absolute addresses for jumps, but Clang decides to pad out the binary
with ~32 KiB of 0s, so that the code *actually* starts at the address
given to "org".

This behavior isn't ideal, since the generated binary violates the
requirement that the initial bootloader fit into the first 512 bytes of
the boot device, but it means that I can use it to bump the 0x55/0xaa
bytes up to their correct locations, while maintaining the ability to
use "jmp" before that point. This isn't portable to NASM, but it's
assembly, so portability is not as much of a concern. Not sure how
GCC/GAS would treat it, though...

It might also be possible to put the bytes in the right locations using
some linker script magic, but it seems like it could be more difficult
to get right -- I may not be experienced with assembly, but I'm even
less experienced with linker scripts, and figuring out exactly what is
going on in the script would likely be much more difficult than what it
took to figure out the current method. Seems support for linker scripts
is pretty new in LLD, too, and still pretty buggy.

Writing a bootloader in the first place isn't ideal. Using GRUB or
Multiboot would probably be easier and more pleasant, but I cannot seem
to create an ISO QEMU will actually boot from. The tutorials on the
internet generally have the user use the old-fashioned BIOS mode for
QEMU, and as a result require GRUB to also support booting through BIOS.
In some cases, this requires downloading a separate GRUB package, since
the default install on EFI systems will try to create EFI images by
default, which won't work for BIOS QEMU.

GRUB isn't available through Homebrew, so I have to build it from
source, at the 2.02 or grub-2.02 tags. This Retina Pro is an EFI system,
so GRUB would normally build for an EFI system, which isn't what the
tutorials require (Following the tutorials exactly isn't strictly
necessary, but it helps isolate mistakes on my end from platform issues
or tutorial issues). It is possible to force GRUB to build a normal BIOS
version, but unfortunately compilation fails due to objconv choking on
some binary or another. This means that BIOS GRUB/QEMU isn't an option.

I tried downloading and using an EFI firmware image for QEMU (OVMF), and
using the default EFI GRUB build to create an ISO, but that also failed
to boot. I have very little experience debugging at this level of
operation, so only the nuclear option is left -- write my own
bootloader!

This is probably going to be far more painful than I initially
anticipate...

    [0]: http://lists.llvm.org/pipermail/llvm-dev/2016-April/098482.html