Add step to canonicalize cs:ip
(Commit message partially copied from last commit)

The BIOS loads the bootloader into linear address 0x7c00. Different
BIOSes use different representations of this, though; most BIOSes use
0x0000:0x7c00, and a few use 0x07c0:0x0000. This necessitates
canonicalizing the values in the code segment register and instruction
pointer register; otherwise, funky things can/will happen (I think).

Canonicalizing cs:ip can be done by performing an absolute long jump to
the next line of code. This sets both %cs and %ip, which allows the code
to continue knowing that certain values will be in those registers.

Adding 0x7c00 to mbr_boot manually is necessary since the bootloader is
loaded at 0x0000:0x7c00 (or equivalently, 0x07c0:0x0000), and the label
represents the offset from the start of the binary. Thus, to jump to the
right place, the program offset has to be added to the label offset.

This may not be necessary if .org didn't cause Clang to insert ~32kb of
0s at the beginning of the binary, or if I could figure out how to use
ORIGIN in linker scripts correctly...

The name "stage_1_prelude" was chosen because this code is intended to
eventually become the stage 1 bootloader, and this canonicalization code
is the first step in getting stage 1 set up. This code (along with the
future relocation code) only needs to run once, so the "prelude" name
felt fitting, since this code precedes the "real" stage 1 bootloader.