Move stage 1 to lower memory before loading starts
Moves the yet-to-be-executed stage 1 code to a lower place in memory.
This is normally used for chainloading, which likely isn't going to be
implemented any time soon, but has the potentially convenient side
effect of making the usable memory contiguous, instead of having some
space, then the bootloader, then more space. Doubt I'll be using all
that space any time soon, but it might be possible to generalize this
code to be able to copy a more complex/memory-hungry bootloader to lower
memory, where not having to deal with a "hole" in available memory would
be nice.

The target address of 0x0060:0x000 was chosen because the OSDev wiki
claimed that that was the "traditional" address the MBR is usually
relocated to when chainloading.

Not copying all 512 bytes of the bootloader code isn't strictly
necessary, but I just felt like it, so I did. The optimization doesn't
consume a significantly greater amount of space (I think), and is still
relatively easy to understand (I hope). Saving a few cycles may not
really matter with hardware these days, but this is a kernel, so might
as well eke out performance where I can.

In any case, I think some math would be necessary to ensure that the CPU
jumps to the actual stage 1 code instead of the relocation code again,
so it can be argued that this is simpler.

Also moved the '!' printing code under the stage_1 label instead of
deleting it because it serves as a convenient way to check the computer
actually gets to the stage 1 code, instead of branching to some
unexpected place.

The cld instruction was added because I'm not sure what the state of the
direction flag is when the computer boots