Limit initial bootloader code to below 0x446
The MBR contains more than just the bootstrap code; in MBR-partitioned
systems, the partition table takes up some space. This places further
restrictions on the size of the bootloader code.

Wikipedia says that the bootstrap code area in a "classical generic MBR"
starts at offset 0 and has a size of 446 bytes. OSDev wiki states that
the MBR bootstrap area has a size of 436 bytes. The field immediately
after the bootloader code is a 10-byte field used for the "unique" disk
ID. Apparently this was never standardized, though, so it can contain
anything -- including the last few bytes of the bootloader code. 446
seems like a reasonable upper bound, then, since chances are nothing is
going to rely on the last 10 bytes.