Remove initial cs:ip canonicalization jmp
I think the way the instruction pointer moves in physical memory is the
same, no matter if the bootloader is loaded at 0x0000:0x7c00 or
0x07c0:0x0000. In addition, because the relocation code explicitly sets
register values, it should behave the same for both values of cs:ip.
This means that the first long jump should not be necessary, since the
relocation code contains another one right after it finishes copying,
which effectively canonicalizes cs:ip anyways.