Set segment registers and stack pointer
Segment register values were set to the value contained by the code
segment register (0x0060) because the CMU OS class writeup for writing a
bootloader [1] says that using the first 64 KiB for most memory
operations allows us to avoid using long/far jmps, which sounds like a
decent reason to me.

Setting %ds/%es isn't strictly necessary, but might as well, since it
makes pretending that we're working in a flat memory space a bit easier.

The idea behind setting %sp to 0xFFF0 is to have the stack begin at the
top of the address range for this particular segment (0x0060), so it has
the most amount of space possible to grow downwards. In addition, with
this value the stack is aligned to 16 bytes, which appears to be the
expected alignment for x86_64 code (and/or code using SSE/related
instructions?) [2].

    [1]: http://www.cs.cmu.edu/~410-s07/p4/p4-boot.pdf
    [2]: https://software.intel.com/en-us/forums/intel-isa-extensions/topic/291241