Disable interrupts during relocation/stack setup
I *think* this is necessary in this case since interrupts cause stuff to
be saved to the stack [1], and the stack has not been set up yet.

Might also be (partially) a relic from the 8088 era? Apparently loading
into ss is supposed turn off interrupts until the end of the next
instruction [2], presumably so ss and sp can be set atomically (to the
rest of the system, at least).

Unfortunately, quite a few 8088s lacked that behavior due to a hardware
bug [3]. Thus, cli/sti were used to ensure that ss and sp would be set
together even with a defective chip. Without those instructions, an
interrupt that occured after ss was set but before sp was set would
effectively cause the computer to write saved values to a random memory
location, and Bad Things (tm) could/would happen.

Either way, interrupts require a functioning stack, and there isn't one
when the bootloader first starts executing.

    [1]: https://books.google.com/books?id=1L7PVOhfUIoC&pg=PA492 :
           In order for processing to continue after the interrupt has
           ended, the at least three values must be saved on the stack:
           the segment and offset of the instruction where execution is
           to be resumed, and the flags register.

    [2]: https://stackoverflow.com/a/39661375 :
             Note that loading ss with mov ...,ss actually turns
             off interrupts for the next instruction. -- fuz

    [3]: https://stackoverflow.com/a/39661375 :
              I know that and under the actual answer I use that
              technique in my comment without CLI. The only caveat is
              that there were a sizable quantity of physical 8088s with
              a bug where interrupts were not turned off until the end
              of the next instruction. So many real bootloaders use
              CLI before setting the stack. -- Michael Petch