Implement basic string printing
Simple loop to read a zero-terminated string and print it using the
BIOS.

Have to pay a code size penalty (for now) for hardcoding the printing of
\r\n after a string is printed, but as long as the number of strings
printed is greater than zero, the total savings will be positive.
Chances are the savings won't be that great, though, since space in a
bootloader is quite limited.

Also created a "print" macro to help with readability and ease of use.
It's nice not having to remember that the address of the string needs to
be loaded into %ds:%si